2605 - code refactoring and unit test cases fix AB#2605 AB#7199
