+++ UKHO.ExchangeSetService.API/UKHO.ExchangeSetService.Common/Helpers/IAzureBlobStorageService.cs
@@ -1,0 +1,12 @@
+﻿using System.Threading;
+using System.Threading.Tasks;
+using UKHO.ExchangeSetService.Common.Models.SalesCatalogue;
+
+namespace UKHO.ExchangeSetService.Common.Helpers
+{
+    public interface IAzureBlobStorageService
+    {        
+        Task<bool> StoreSaleCatalogueServiceResponseAsync(string containerName, string batchId, SalesCatalogueProductResponse salesCatalogueResponse, string callBackUri, string correlationId, CancellationToken cancellationToken);
+        Task<SalesCatalogueProductResponse> DownloadSalesCatalogueResponse(string scsResponseUri);
+    }
+}
