+++ UKHO.ExchangeSetService.API/UKHO.ExchangeSetService.FulfilmentService/Program.cs
@@ -90,6 +90,7 @@ namespace UKHO.ExchangeSetService.FulfilmentService
                  services.AddScoped<IEssFulfilmentStorageConfiguration, EssFulfilmentStorageConfiguration>();
                  services.AddScoped<ISalesCatalogueStorageService, SalesCatalogueStorageService>();
                  services.AddScoped<IFulfilmentDataService, FulfilmentDataService>();
+                 services.AddScoped<IAzureBlobStorageService, AzureBlobStorageService>();
                  services.AddScoped<IAzureBlobStorageClient, AzureBlobStorageClient>();
                  services.AddScoped<IAzureMessageQueueHelper, AzureMessageQueueHelper>();
                  services.AddHttpClient<IFileShareServiceClient, FileShareServiceClient>(client =>
